refactor: Menu와 MenuProduct 패키지 병합
