+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -6,9 +6,9 @@ import kitchenpos.menu.MenuName;
 import kitchenpos.menu.MenuPrice;
 import kitchenpos.menu.MenuProductValidator;
 import kitchenpos.menu.MenuRepository;
+import kitchenpos.menu.application.request.MenuProductRequest;
 import kitchenpos.menu.application.request.MenuRequest;
 import kitchenpos.menugroup.MenuGroupRepository;
-import kitchenpos.menuproduct.request.MenuProductRequest;
 import kitchenpos.product.Product;
 import kitchenpos.product.ProductName;
 import kitchenpos.product.ProductPrice;
